Desain Admin dan User

+++ app/Http/Controllers/Menu/BeritaController.php
@@ -5,20 +5,40 @@ namespace App\Http\Controllers\Menu;
 use App\Http\Controllers\Controller;
 use App\Models\Berita;
 use Illuminate\Http\Request;
-use Illuminate\Support\Carbon;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Validator;
+use Illuminate\Support\Carbon;
 use Illuminate\Support\Str;
+use Yajra\DataTables\Facades\DataTables;
 
 class BeritaController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
-        $Berita = Berita::all();
-        return view('menu/berita/list', compact('Berita'));
+        if ($request->ajax()) {
+            $model = Berita::query()->select('id', 'judul');
+            if ($request->has('order') && !empty($request->input('order'))) {
+                $order = $request->input('order')[0];
+                $columnIndex = $order['column'];
+                $columnName = $request->input('columns')[$columnIndex]['data'];
+                $columnDirection = $order['dir'];
+                $model->orderBy($columnName, $columnDirection);
+            } else {
+                $model->latest();
+            }
+            return DataTables::eloquent($model)
+                ->addIndexColumn()
+                ->addColumn('action', function ($row) {
+                    return '<a class="w-32-px h-32-px bg-primary-light text-primary-600 rounded-circle d-inline-flex align-items-center justify-content-center" href="' . route('kelola-berita.show', $row->id) . '"> <iconify-icon icon="iconamoon:eye-light"></iconify-icon> </a> <a class="w-32-px h-32-px bg-success-focus text-success-main rounded-circle d-inline-flex align-items-center justify-content-center" href="' . route('kelola-berita.edit', $row->id) . '"> <iconify-icon icon="lucide:edit"></iconify-icon> </a> <a class="w-32-px h-32-px bg-danger-focus text-danger-main rounded-circle d-inline-flex align-items-center justify-content-center" style="cursor:pointer" onclick="hapus(\'' . $row->id . '\')"> <iconify-icon icon="mingcute:delete-2-line"></iconify-icon> </a>';
+                })
+                ->rawColumns(['action'])
+                ->removeColumn('id')
+                ->toJson(true);
+        }
+        return view('menu/berita/list');
     }
 
     /**
@@ -36,8 +56,9 @@ class BeritaController extends Controller
     {
         $rules = [
             'judul-berita' => 'required|string|max:255',
-            'deskripsi-berita' => 'required|string',
             'foto-berita' => 'required|image|mimes:jpeg,jpg,png|max:3072',
+            'kutipan-berita' => 'required|string|max:255',
+            'deskripsi-berita' => 'required|string',
         ];
         $validator = Validator::make($request->all(), $rules);
         if ($validator->fails()) {
@@ -45,8 +66,9 @@ class BeritaController extends Controller
         }
         $validator = $validator->validated();
         $data = [
-            'judul' => $validator['judul-berita'],
-            'deskripsi' => $validator['deskripsi-berita'],
+            'judul' => trim($validator['judul-berita']),
+            'kutipan' => trim($validator['kutipan-berita']),
+            'deskripsi' => trim($validator['deskripsi-berita']),
         ];
         $today = Carbon::now()->setTimezone('Asia/Jakarta')->format('YmdHis');
         $converted = Str::lower($validator['judul-berita']);
@@ -60,7 +82,7 @@ class BeritaController extends Controller
 
         Berita::create($data);
 
-        return redirect()->route('kelola-berita.index')->with(['message' => 'sukses menambahkan data.', 'isActive' => true, 'hasError' => false]);
+        return redirect()->route('kelola-berita.index')->with(['message' => 'sukses menambahkan data.']);
     }
 
     /**
@@ -88,8 +110,9 @@ class BeritaController extends Controller
     {
         $rules = [
             'judul-berita' => 'required|string|max:255',
-            'deskripsi-berita' => 'required|string',
             'foto-berita' => 'nullable|image|mimes:jpeg,jpg,png|max:3072',
+            'kutipan-berita' => 'required|string|max:255',
+            'deskripsi-berita' => 'required|string',
         ];
         $validator = Validator::make($request->all(), $rules);
         if ($validator->fails()) {
@@ -97,8 +120,9 @@ class BeritaController extends Controller
         }
         $validator = $validator->validated();
         $data = [
-            'judul' => $validator['judul-berita'],
-            'deskripsi' => $validator['deskripsi-berita'],
+            'judul' => trim($validator['judul-berita']),
+            'kutipan' => trim($validator['kutipan-berita']),
+            'deskripsi' => trim($validator['deskripsi-berita']),
         ];
         $today = Carbon::now()->setTimezone('Asia/Jakarta')->format('YmdHis');
         $converted = Str::lower($validator['judul-berita']);
@@ -116,20 +140,32 @@ class BeritaController extends Controller
         $getBeritaData->update($data);
 
         return redirect()->route('kelola-berita.index')->with(['message' => 'sukses mengubah data.', 'isActive' => true, 'hasError' => false]);
-
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $id)
+    public function destroy(Request $request, string $id)
     {
-        $Berita = Berita::findOrFail($id);
-        if (!empty($Berita->file_foto)) {
-            Storage::delete($Berita->file_foto);
-        }
-        $Berita->delete();
-        return redirect()->route('kelola-berita.index')->with('message', 'sukses menghapus data.');
+        if ($request->ajax()) {
+            try {
+                $Berita = Berita::findOrFail($id);
+                if (!empty($Berita->file_foto)) {
+                    Storage::delete($Berita->file_foto);
+                }
+                $Berita->delete();
 
+                return response([
+                    'status' => 'success',
+                    'data' => 'sukses menghapus data.'
+                ]);
+            } catch (\Exception $e) {
+                return response([
+                    'status' => 'error',
+                    'data' => $e->getMessage()
+                ], 500);
+            }
+        }
+        return redirect()->back();
     }
 }
